evaluator-gcp-etl3/requirements.txt
google-cloud-bigquery
google-cloud-spanner
google-cloud-dataflow
